ui modularization
